Changed key type in CachedAnnotationTypes dictionary
The key type of the `CachedAnnotationTypes` dictionary in the `TAnnotationCaching` class has been changed from `Type` to `string`. This change has been reflected in the `HasAnnotationAttr` property of the `ISetting` interface, where `typeof(T).FullName!` is now used to retrieve a value from the dictionary instead of `typeof(T)`. Similarly, when adding a new entry to the `CachedAnnotationTypes` dictionary, `typeof(T).FullName!` is used as the key instead of `typeof(T)`.